# market/src/main/resources/api-docs.yaml
openapi: 3.1.1
info:
  title: Market Service API
  version: 1.0.0
paths:
...
  /companies/{companyId}:
    get:
      summary: Retrieves a company by its unique ID.
      operationId: getCompany
      parameters: # list of parameters, must define each starting with -
        - $ref: "#/components/parameters/CompanyId"
      responses:
        "200":
          description: A company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "400":
          $ref: "#/components/responses/InvalidCompanyId"
        "404":
          $ref: "#/components/responses/CompanyNotFound"
    delete:
      summary: Remove a company from the market.
      operationId: removeCompanyFromMarket
      parameters:
        - $ref: "#/components/parameters/CompanyId"
      responses:
        "204":
          description: Company removed from market.
        "400":
          $ref: "#/components/responses/InvalidCompanyId"
        "404":
          $ref: "#/components/responses/CompanyNotFound"

  # Note on use of path parameters vs. query
  # query is used to filter, search, or modify request to resource
    # e.g. GET /companies?id=123 - search for companies with ID 123
  # path represents a specific resource, use when directly addressing
    # e.g. POST /companies/123/buy - buy shares from company 123
...
components:
...
  parameters:
    CompanyId:
      name: companyId
      in: path # location of parameter - e.g. companyId located in url path
      schema: # define structure of parameter
        type: int
      description: Company ID
      required: true # path parameters always required
  # define common HTTP response blocks
  responses:
    InvalidCompanyId:
      description: Bad request. Company ID in an invalid format.
      content:
        application/json:
          # response must follow structure defined in Error schema
          schema:
            $ref: "#/components/schemas/Error"
    CompanyNotFound:
      description: Company with specified ID not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  # defining common HTTP request bodies
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...
    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sharePrice:
          type: number
        shareMultiplier:
          type: number
        numShares:
          type: integer
      required:
        - id
        - name
        - sharePrice
        - numShares